move style file, moved linegroupstyle into table_style
