+++ planner/python/classes/elements/entry_table.py
@@ -31,8 +31,8 @@ import svgwrite.shapes
 import svgwrite.text
 
 from classes.constants.dims import PlannerDims as Dims
-from classes.constants.style import PlannerColors as Colors
-from classes.constants.style import PlannerFontStyle as Font
+from classes.style.style import PlannerColors as Colors
+from classes.style.style import PlannerFontStyle as Font
 from classes.constants.strings import PlannerStrings as Strings
 from classes.constants.error_strings import ErrorStrings as Err
