+++ planner/python/classes/elements/table_rows.py
@@ -30,8 +30,8 @@ import svgwrite
 
 from classes.constants.dims import PlannerDims as Dims
 from classes.constants.error_strings import ErrorStrings as Err
-from classes.constants.style import PlannerFontStyle as Font
-from classes.constants.style import PlannerColors as Colors
+from classes.style.style import PlannerFontStyle as Font
+from classes.style.style import PlannerColors as Colors
 from utils.utils import PlannerUtils as Utils
 
 from classes.elements.rows import Rows
