+++ planner/python/classes/page_entries/week_entry.py
@@ -30,8 +30,8 @@ import svgwrite.container
 import svgwrite.text
 
 from classes.constants.strings import PlannerStrings as Strings
-from classes.constants.style import PlannerColors as Colors
-from classes.constants.style import PlannerFontStyle as Font
+from classes.style.style import PlannerColors as Colors
+from classes.style.style import PlannerFontStyle as Font
 
 from classes.elements.entry_group import EntryRow
 from classes.elements.entry_table import EntryTable
