+++ planner/python/utils/utils.py
@@ -1,6 +1,6 @@
 import svgwrite.container
-from classes.constants.style import PlannerFontStyle as Font
-from classes.constants.style import PlannerColors as Colors
+from classes.style.style import PlannerFontStyle as Font
+from classes.style.style import PlannerColors as Colors
 from classes.constants.error_strings import ErrorStrings as Err
 from classes.constants.dims import PlannerDims as Dims
